server/requirements.txt
requests
numpy
rembg
Pillow
scikit-image
